ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.8.0
pydantic==2.5.3
pydantic-settings==2.1.0
confluent-kafka==2.3.0
redis==5.0.1
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
httpx==0.26.0
python-json-logger==2.0.7
structlog==24.1.0
numpy==1.26.3
scikit-learn==1.3.2
requests==2.31.00
python-dotenv==1.0.0
